Move modules for backend API under `backend` module
To group all backend-related stuff in one place.
