+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Rename `ChannelKind` into just `Channel`.
 - Rename `channels` module into `replicon_channels`.
+- Move `replicon_server`, `replicon_client`, `connected_client` and `replicon_channels` under `backend` module to group all backend-related API.
 - All methods with `Into<Channel>` now just accept `Channel`.
 - Use `usize` for channel ID. Backends now decide how many channels user can create.
 - Don't insert `ClientVisibility` at all if `ServerPlugin::visibility_policy` is set to `VisibilityPolicy::All`. Previously all calls were just no-op.
